requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
